+++ x/perp/v2/keeper/clearing_house.go
@@ -725,7 +725,15 @@ func (k Keeper) ClosePosition(ctx sdk.Context, pair asset.Pair, traderAddr sdk.A
 	}
 
 	if positionResp.BadDebt.IsPositive() {
-		return nil, fmt.Errorf("underwater position")
+		if err = k.realizeBadDebt(
+			ctx,
+			market,
+			positionResp.BadDebt.RoundInt(),
+		); err != nil {
+			return nil, err
+		}
+
+		return positionResp, nil
 	}
 
 	if err = k.afterPositionUpdate(